fix prefered to float

+++ core/api-server/api/graphql/schemas/prefered-schema.js
@@ -9,7 +9,7 @@ type ReturnList {
   pipelineName: String
   priority: Int
   maxExceeded: Boolean
-  entranceTime: Int
+  entranceTime: Float
   tags: [String]
 }
